requirements.txt
google-api-python-client==2.91.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pymongo==4.4.0
requests==2.31.0
pandas~=2.0.3
python-dotenv~=1.0.0